requirements.txt
django
djangorestframework
cloudinary
python-dotenv
dj-rest-auth
django-allauth
pillow
djangorestframework-simplejwt